Remove final slash which produces issues with fastapi
 I am trying to deploy gradio ChatInterface to a server and streaming only works locally, but not on a server like google cloud run or fly.io. Without streaming it works like a charm. 

It seems to be due to this redirect: 


2023-08-08T09:47:03.430 app[d891699f69ed98] ams [info] 2023-08-08 09:47:03,431 DEBUG MainThread httpcore.http11.atrace:85 receive_response_headers.complete return_value=(b'HTTP/1.1', 301, b'Moved Permanently', [(b'location', b'https://xxx/gradio-chat/api/predict'), (b'server', b'Fly/49bc237b (2023-08-04)'), (b'via', b'1.1 fly.io'), (b'fly-request-id', b'01H7A8NWW406RA4E76MK913315-ams'), (b'content-length', b'0'), (b'date', b'Tue, 08 Aug 2023 09:47:02 GMT')])


Which does not happen locally and afterwards this redirect the websocket connection is closed.

Can i prevent this redirect from happening? I guess due to (b'content-length', b'0') the error: Expecting value: line 1 column 1 (char 0) happens, but this is of course only a side effect of the redirect.